+++ visualizer/Animator.java
@@ -95,8 +95,8 @@ public class Animator extends PApplet {
 
         void setScale() {
             // update sizes
-            this.solutionHeight = this.solution.getHeight();
-            this.solutionWidth = this.solution.getWidth();
+            this.solutionHeight = (int) this.solution.getHeight();
+            this.solutionWidth = (int) this.solution.getWidth();
 
             if (this.solutionWidth > this.solutionHeight) {
                 drawWidth = maxSize;
